visualize final results - fixed and improved
Corrected notebook file is uploaded instead of old one (mistakenly attached in previous commit).
Furthermore, now the notebook also creates a plot for the misclassified labels of the SA task, between different tones.
That plot is now attached, too, in addition to all the other plots the notebook creates that were attached in previous commit.